deepen: fold OpenRouter Credit Monitor into Slack daily-report
`lib/ai/openrouter-credit-monitor.ts` (161 LOC) exported two functions —
`getOpenRouterCreditStatus` (HTTP fetch to OpenRouter's `/api/v1/auth/key`)
and `formatCreditStatusForSlack` (Slack-tone block rendering) — plus the
`CreditStatus` type. Both functions had exactly one production caller:
`lib/slack/daily-report-handler.ts:generateHourlySummary`. No tests
existed at the deleted module's **interface**.

Per LANGUAGE.md, one **adapter** is hypothetical, not a real **seam**.
Both functions live exclusively for the "OpenRouter Credits" block
rendered in the hourly Slack report. The seam between them and the
Slack handler added zero **leverage**: a Slack-tone formatter does not
belong in `lib/ai/`, and the HTTP fetch is structurally identical to
the other Slack-specific telemetry helpers that already live next to
it in `daily-report-handler.ts`.

**Leverage**: the hourly Slack report now reads top-to-bottom for the
credit-status block — no jump to `lib/ai/` to read what the
**interface** actually does. The two functions are filed alongside the
other private telemetry helpers (queue status, filing breakdown, AI
cost rollup) in the file that owns them.

**Locality**: changes to the credit-warning threshold, the OpenRouter
endpoint, or the Slack-tone rendering now concentrate in one file
instead of two. The header comment in the inlined block makes the
provenance explicit so future maintainers see why the HTTP fetch lives
under `lib/slack/` rather than `lib/ai/`.

**Dependency category**: True external (deepening.md §4). The deepened
caller depends on OpenRouter's HTTP API. The caller already had no
**adapter** seam for that dependency — the fetch is inline. Inlining
the credit-monitor preserves that shape; tests that mock OpenRouter
fetches will mock at `fetch` as they already do.

**Tests deleted at the deepened module's interface**: zero — the
deleted **module** had no dedicated test file.
**Tests added at the deepened module's interface**: zero — Slack-handler
behaviour is already covered through its existing report-generation
seam.

The lone comment reference in `__tests__/security/budget-manipulation.test.ts`
is updated to point at the new location.

**ADRs respected**:
- ADR-0001 / ADR-0002 — same shape: collapse a one-caller extraction
  whose seam was hypothetical, not real.
- ADR-0006 (Financial Content Gate stays modular) — does not apply.
  ADR-0006's load-bearing reason was the Gate's 36-test surface that
  integration tests at its caller seam did not yet cover. Here, the
  deleted **module** had zero tests, so collapsing the seam costs
  nothing.
